% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_na_percentage.R
\name{calculate_na_percentage}
\alias{calculate_na_percentage}
\title{Calculate NA Percentage}
\usage{
calculate_na_percentage(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric value representing the percentage of NA values in the vector.
}
\description{
This function calculates the percentage of NA values in a given vector. It also includes a margin of 10\%.
}
