% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed_package.R
\name{check_installed_package}
\alias{check_installed_package}
\title{check_installed_package}
\usage{
check_installed_package(package_name, check = FALSE)
}
\arguments{
\item{package_name}{the name of the package (quoted)}

\item{check}{the function should work as a boolean operator}
}
\value{
Boolean value whether package is installed.
}
\description{
Check if a package is installed. If not,
throw an error message
}
\examples{
  check_installed_package("dplyr")

}
