% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentation.R
\name{create_documentation}
\alias{create_documentation}
\title{Create Documentation}
\usage{
create_documentation(df, doc_file_name)
}
\arguments{
\item{df}{The dataframe for which documentation is to be created.}

\item{doc_file_name}{Name of the documentation file to create.}
}
\value{
A table containing the documentation.
}
\description{
This function creates documentation for a given dataframe.
}
