% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_class.R
\name{first_class}
\alias{first_class}
\title{Get the first class of an object}
\usage{
first_class(obj)
}
\arguments{
\item{obj}{The object or vector whose class is to be found.}
}
\value{
The first class of the object or vector.
}
\description{
This function returns the first class of an object or vector.
}
\examples{
# Create a vector of different classes
my_vector <- c(1, "two", 3.0)
# Get the first class of the vector
first_class(my_vector)
}
