% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixencoding.R
\name{fixencoding}
\alias{fixencoding}
\title{Fix encoding}
\usage{
fixencoding(df, originalEncoding = "latin1")
}
\arguments{
\item{df}{Dataframe}

\item{originalEncoding}{The encoding to apply}
}
\value{
Dataframe
}
\description{
a generic function that can read an R object into an RDS, fst and CSV.
First determine the extension and then apply the appropriate read function
}
