% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Get Distribution}
\usage{
get_distribution(data_vector)
}
\arguments{
\item{data_vector}{Column to check.}
}
\description{
This function gets the distribution for numeric columns.
}
\seealso{
Other assertions: 
\code{\link{get_ratio}()},
\code{\link{is_field_subset}()}

Other tests: 
\code{\link{get_ratio}()},
\code{\link{is_field_subset}()}
}
\concept{assertions}
\concept{tests}
