% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_last_modified.R
\name{get_last_modifed_date}
\alias{get_last_modifed_date}
\title{get_last_modified_date}
\usage{
get_last_modifed_date(file_path)
}
\arguments{
\item{file_path}{Path to the file.}
}
\value{
Date that file was last modified
}
\description{
get_last_modified_date
}
