% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{get_last_n_commit_date}
\alias{get_last_n_commit_date}
\title{Retrieve the last n commit dates of a file}
\usage{
get_last_n_commit_date(filepath, n = 1)
}
\arguments{
\item{filepath}{Path to file}

\item{n}{Integer number to look back in the git log}
}
\description{
Retrieve the last n commit dates of a file
}
