% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{get_last_n_commit_info}
\alias{get_last_n_commit_info}
\title{Retrieve the last n commits of a file}
\usage{
get_last_n_commit_info(filepath, type = "", n = 1)
}
\arguments{
\item{filepath}{Path to file}

\item{type}{The type of commit, according to commit style guide}

\item{n}{Integer number to look back in the git log}
}
\description{
Retrieve the last n commits of a file
}
\examples{
\dontrun{
# get the last three commits which were bug fixes
get_last_n_commit_info("01. Inlezen/Inlezen Aanmeldingen.R", type = "fix", n = 3)
}
}
