% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratio.R
\name{get_ratio}
\alias{get_ratio}
\title{get_ratio}
\usage{
get_ratio(data_vector)
}
\arguments{
\item{data_vector}{column to check}
}
\description{
Get percentages for factor columns
}
\seealso{
Other assertions: 
\code{\link{get_distribution}()},
\code{\link{is_field_subset}()}

Other tests: 
\code{\link{get_distribution}()},
\code{\link{is_field_subset}()}
}
\concept{assertions}
\concept{tests}
