% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_file.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Get recent file}
\usage{
get_recent_file(path, match, date_type = "modified")
}
\arguments{
\item{path}{The path to search for the file}

\item{match}{The search term matched in the file name}

\item{date_type}{The way to find the recent file
date_type = "modified" is based on customization, date_type = "filename_ymd"
is based on file name.}
}
\value{
The most recent file.
}
\description{
Is a wrapper around get_recent_file_date_filename_ymd en
get_recent_file_date_modified and retrieves the most recent version of a
file based on naming or date modified.
}
\seealso{
Other Get recent files: 
\code{\link{get_recent_file_date_filename_ymd}()},
\code{\link{get_recent_file_date_modified}()}
}
\concept{Get recent files}
