% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_file_date_filename_ymd.R
\name{get_recent_file_date_filename_ymd}
\alias{get_recent_file_date_filename_ymd}
\title{Get recent file date filename ymd}
\usage{
get_recent_file_date_filename_ymd(path, match)
}
\arguments{
\item{path}{The path to search for the file.}

\item{match}{The search term to match in the file name.}
}
\value{
The most recent file
}
\description{
This function determines the path of the most recent version of a file in
a folder. Sorting is determined by file name
where it is a condition that the filename starts with ymd encoding.
}
\seealso{
Other Get recent files: 
\code{\link{get_recent_file_date_modified}()},
\code{\link{get_recent_file}()}
}
\concept{Get recent files}
