% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_file_date_modified.R
\name{get_recent_file_date_modified}
\alias{get_recent_file_date_modified}
\title{Get recent file date modified
This function determines the path of the most recent version of a file in
a folder. The sorting is determined based on the date of the last
change.}
\usage{
get_recent_file_date_modified(path, match, echo = TRUE)
}
\arguments{
\item{path}{The path to search for the file.}

\item{match}{The search term to match in the file name.}

\item{echo}{Print the date the file was last modified
in the console.}
}
\value{
The most recent file.
}
\description{
Get recent file date modified
This function determines the path of the most recent version of a file in
a folder. The sorting is determined based on the date of the last
change.
}
\seealso{
Other Get recent files: 
\code{\link{get_recent_file_date_filename_ymd}()},
\code{\link{get_recent_file}()}
}
\concept{Get recent files}
