% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_field_subset.R
\name{is_field_subset}
\alias{is_field_subset}
\title{Check if Field is Subset}
\usage{
is_field_subset(field_name, df, column_names, column_types)
}
\arguments{
\item{field_name}{Name of the field to check.}

\item{df}{Dataframe to check.}

\item{column_names}{Set that the field is part of.}

\item{column_types}{Types that belong to the column names.}
}
\description{
This function checks if a field is a subset.
We define a subset as a character field that contains fewer than 20 unique values.
}
\seealso{
Other tests: 
\code{\link{get_distribution}()},
\code{\link{get_ratio}()}

Other assertions: 
\code{\link{get_distribution}()},
\code{\link{get_ratio}()}
}
\concept{assertions}
\concept{tests}
