% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique_identifier.R
\name{is_unique_identifier}
\alias{is_unique_identifier}
\title{Is Unique Identifier}
\usage{
is_unique_identifier(field_name, df)
}
\arguments{
\item{field_name}{The name that is being tested for uniqueness.}

\item{df}{Dataframe where the unique identifier is being searched for.}
}
\description{
This function checks if a field is a unique identifier.
}
