% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_numeric.R
\name{max_numeric}
\alias{max_numeric}
\title{Maximum Numeric}
\usage{
max_numeric(x)
}
\arguments{
\item{x}{Vector of which maximum is wanted.}
}
\description{
This function finds the maximum of a numeric vector.
If the vector is not numeric, it returns NA.
}
