% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_numeric.R
\name{min_numeric}
\alias{min_numeric}
\title{Minimum Numeric}
\usage{
min_numeric(x)
}
\arguments{
\item{x}{Vector of which minimum is wanted.}
}
\description{
This function finds the minimum of a numeric vector.
If the vector is not numeric, it returns NA.
}
