% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_last_modified.R
\name{print_last_modified}
\alias{print_last_modified}
\title{Print date the file was last modified}
\usage{
print_last_modified(path)
}
\arguments{
\item{path}{The file path}
}
\value{
message
}
\description{
This function prints the date a file was last modified
}
\examples{
print_last_modified(readr::readr_example("mtcars.csv"))

}
