% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_string_vector.R
\name{random_string_vector}
\alias{random_string_vector}
\title{Generate a random string vector}
\usage{
random_string_vector(
  n = 500,
  length = 6,
  characters = c(letters, LETTERS, 0:9)
)
}
\arguments{
\item{n}{The number of items in the vector. Default is set to 500.}

\item{length}{the number of characters in a string. Default is set to 6.}

\item{characters}{A vector containing the characters to include. Default is all lowercase, all, uppercase letters and all numbers.}
}
\description{
Generate a random string vector
}
