% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheet.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read Excel Allsheets}
\usage{
read_excel_allsheets(filename)
}
\arguments{
\item{filename}{Name of Excel file}
}
\value{
Dataframe
}
\description{
Read in all sheets in an Excel file.
}
\examples{
read_excel_allsheets(readxl::readxl_example("clippy.xls"))

}
