% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_session_info_to_file.R
\name{save_session_info_to_file}
\alias{save_session_info_to_file}
\title{Save session info to a file}
\usage{
save_session_info_to_file(path)
}
\arguments{
\item{path}{The directory path where the session info file will be saved.}
}
\value{
A .txt file containing session info, saved at the specified path.
}
\description{
Stores session info in a .txt file.
}
