% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_sessioninfo.R
\name{save_sessioninfo}
\alias{save_sessioninfo}
\title{Save session info}
\usage{
save_sessioninfo(path)
}
\arguments{
\item{path}{The path to search for the file.}
}
\value{
.txt file containing the session info.
}
\description{
Stores session info.
}
