% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_file_date.R
\name{set_file_date}
\alias{set_file_date}
\title{Set file data}
\usage{
set_file_date(data, prefix, stage)
}
\arguments{
\item{data}{Data frame.}

\item{prefix}{Textstring.}

\item{stage}{Textstring.}
}
\value{
Data frame.
}
\description{
Add datet to file name.
}
