% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_read_delim.R
\name{unzip_read_delim}
\alias{unzip_read_delim}
\title{Extract a .zip archive and read in with read_delim}
\usage{
unzip_read_delim(zip_path, filename = NULL, ...)
}
\arguments{
\item{zip_path}{The file path of the .zip archive}

\item{filename}{OPTIONAL: The file name of the file in the .zip archive to be read. This parameter can be left empty if there is only 1 file in the archive.}

\item{...}{arguments to the readr::read_delim function. see: \link[readr]{read_delim}}
}
\value{
Dataframe
}
\description{
Extract a .zip archive and read it in with readr's read_delim function. The file is extracted to a temporary location, and then deleted after reading it.
}
\examples{
unzip_read_delim(readr::readr_example("mtcars.csv.zip"))

}
