% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files_pattern.R
\name{move_files_pattern}
\alias{move_files_pattern}
\title{move files pattern}
\usage{
move_files_pattern(Folder_origin, Folder_dest, pattern, recursive = FALSE)
}
\arguments{
\item{Folder_origin}{Source folder.}

\item{Folder_dest}{Destination folder.}

\item{pattern}{Pattern to match files in source folder on.}

\item{recursive}{Default: FALSE. Whether to use recursive search in directory.}
}
\value{
message
}
\description{
Move files in a directory based on regular expression
}
